CampBX: always use BigDecimal instead of double/Double to maintain precision

+++ xchange-campbx/src/main/java/com/xeiam/xchange/campbx/api/model/OrderBook.java
@@ -24,6 +24,7 @@ package com.xeiam.xchange.campbx.api.model;
 
 import org.codehaus.jackson.annotate.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -35,18 +36,18 @@ import java.util.List;
 public class OrderBook {
 
   @JsonProperty("Bids")
-  private List<List<Double>> bids;
+  private List<List<BigDecimal>> bids;
   @JsonProperty("Asks")
-  private List<List<Double>> asks;
+  private List<List<BigDecimal>> asks;
 
   /** (price, amount) */
-  public List<List<Double>> getBids() {
+  public List<List<BigDecimal>> getBids() {
 
     return bids;
   }
 
   /** (price, amount) */
-  public List<List<Double>> getAsks() {
+  public List<List<BigDecimal>> getAsks() {
 
     return asks;
   }
